requirements.txt
anthropic>=0.25.0
openai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
flask>=3.0.0
waitress>=3.0.0
